Add virtio-vsock device
This is developed against v5 of the Linux AF_VSOCK virtio patches
http://thread.gmane.org/gmane.linux.kernel.virtualization/27455
and v4 of the spec proposal + comments in the thread
http://thread.gmane.org/gmane.comp.emulators.virtio.devel/1062 and v5 of the
spec proposal http://thread.gmane.org/gmane.comp.emulators.virtio.devel/1092

The guest vsock connections are proxied into OSX host Unix domain sockets in
a directory configured with the (mandatory) "path=" option.

To open a listening socket on the OSX end create Unix domain socket with a name
in the form CCCCCCCC.PPPPPPPP where CCCCCCCC is the host CID, in hex
zero-padded to 8 characters (so always 00000002) and PPPPPPPP is the port
number (also in hex zero-padded to 8 characters).

To connect from OSX to a listening port in the guest open the "connect" socket
and write "GGGGGGGG.PPPPPPPP\n", where GGGGGGG is the zero-padded hex guest CID
and PPPPPPPP is the port number. The guest CID is configured by the (mandatory)
"guest_cid" option.

A configurable set of ports can also be forwarded ("guest_forwards=" option
which is a colon separated list of port numbers), in which case for each port a
"GGGGGGGG.PPPPPPPP" Unix domain socket will be created which when opened
forwards to that socket. For convenience a symlink "guest.PPPPPPPP" is also
created.

The virtio-protocol requires buffering, however whenever possible we try to
read/write directly from/to to the guest buffers. However is this would block
then we copy into a buffer. When buffering we do not update the peer's notion
of our forward count until the buffer is completely drained, in the hopes of
getting back into the direct read/write mode ASAP rather than have the buffer
being continually topped up. If this turns out to introduce unwanted burstiness
then this can be revisited. The buffers are 32K, which matches the default
buffer in go's io.Copy which is used by vsudd inside the guest to proxy to the
docker socket.

Currently there is a hard coded limit of open connections (1024) and forwards
(4) and arrays are used rather than e.g. lists/hashes, which could obviously be
improved.

Host port numbers are "allocated" as array index + 65536 (which puts them above
the TCP port numbers defined in /etc/services and therefore out of the way of
any listening sockets which may be reusing that number space)

Some (very simple) tests are added to v1/tests/cases/network/vsock/ but these
are not wired up to "make test".

As a protocol wrinkle we only close the connection on an explicit OP_SHUTDOWN
flags == 3 (both RX and TX shutdown), rather then when the union of all past
OP_SHUTDOWN's == 3. With the v5 Linux patches the following sequence of
shutdowns is received from the guest):
1) OP_SHUTDOWN flags == VIRTIO_VSOCK_FLAG_SHUTDOWN_RX
2) OP_SHUTDOWN flags == VIRTIO_VSOCK_FLAG_SHUTDOWN_TX
3) OP_SHUTDOWN flags == VIRTIO_VSOCK_FLAG_SHUTDOWN_TX|VIRTIO_VSOCK_FLAG_SHUTDOWN_RX

If we shutdown at #2 (since both directions are down) and then #3 would be to a
non-existent socket (or, worse, an active one if it was reused). It's not clear
if this is expected f/e behaviour or a bug in the Linux patches.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>