vsock: defer replies generated by the tx thread to the rx thread via a ring.
Currently we send the replies synchronously from the tx thread by spinning
waiting for an rx descriptor, however this results in a deadlock with the
frontend:

A transmitting process on the guest is waiting for a tx slot to become
available, while holding the AF_VSOCK lock.

The rx workqueue which should be replenishing the ring is blocked waiting for
AF_VSOCK.

The rx ring is empty and therefore the backend tx thread will spin and will not
free up any tx slots.

We now have a deadlock because the transmitting process is waiting for those
slots.

To break the deadlock instead push replies onto a ring (internal to the
backend) and have them sent from the rx thread instead. Since the tx thread no
longer needs to spin it can carry on processing and free up tx slots, avoiding
the deadlock.

I initially made the reply_ring have the same number of entries as the ring,
but observed this being exhausted(!) so have doubled it.

I think this is a symptom of a bug in the underlying virtio-vsock protocol
specification, since there appears to be no limit on the number of requests a
guest can tx which require replies, and consequently the backend can either
buffer an arbitrary number of replies or risk deadlocking. This change seems to
serve well enough as a band aid in my tests while I bring this up upstream.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>