vsock: avoid sending CREDIT_UPDATE replies over the reply ring
This ring tends to fill up resulting in an assert (the other option being a
deadlock).

To mitigate this somewhat we can signal the need to send a CREDIT_UPDATE reply
via a bit on each socket struct instead of using up a ring slot since all the
information needed to construct the response is in the socket struct anyway.
Since most of the reply traffic ought to be CREDIT_UPDATE replies this ought to
releive the presure considerably. It should also end up consolidating multiple
back to back CREDIT_UPDATES.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>