vsock: add a generation counter to source port allocation
Currently the source port for a host initiated connection is "allocated" using
the offset of the struct pci_vtsock_sock datastructure in the socks array (plus
a fixed offset). This opens up the posibility of a source port being reused in
fairly short order, which in turn can mean that an OP_SHUTDOWN sent by the peer
as part of a previous connection can disrupt a new connection.

This is fixed in v6 of the protocol proposal, by requiring an OP_RST as a
response to any OP_SHUTDOWN which results in connection teardown however the
currently posted frontend RFC patches do not yet implement this.

To mitigate this issue until the frontend patches are updated add a 16-bit
per-struct pci_vtsock_sock generation counter and include this in the lower 16
bits of the assigned port number moving the array offset (+1 to avoid assigning
port 0) to the top 16 bits. An old/delayed OP_SHUTDOWN now needs to be 2^16
recycles of a given struct behind to cause the problem described above.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>